+++ Magic/src/main/java/com/elmakers/mine/bukkit/maps/URLMap.java
@@ -288,7 +288,7 @@ public class URLMap extends MapRenderer implements com.elmakers.mine.bukkit.api.
         boolean matched = false;
         while (!matched && retry < maxIds) {
             MapView newView = Bukkit.createMap(world);
-            short newId = CompatibilityLib.getDeprecatedUtils().getMapId(newView);
+            int newId = CompatibilityLib.getDeprecatedUtils().getMapId(newView);
             matched = newId == id;
             if (newId < 0 || newId > id) break;
             retry++;
